finish fibonacci
to fulfill the rules of Fibonacci, I added the predecessors together and stored them in $t3, and shifted the values for the next iteration of the loop. Then by rule, I have to decrease the original input by one and restart the loop. I have to make sure it doesn't decrease lower than 0 otherwise it would not be a Fibonacci sequence.